Incorporate provider configuration with backend
This enables a more flexible provider configuration as the docs
already presuppose editing the env backend.tf files. This allows
for other types of provider configuration (for example using
roles rather than named profiles), default tags, "allowed" account
protection etc.

To support this use case the aws_ vars have been removed and need
to be set directly in backend.tf.

This is intended to be backwards compatible i.e. the docs do not
need to be updated to support these changes.